requirements.txt
asgiref==3.7.2
Django==5.0.1
dnspython==2.4.2
packaging==23.2
plotly==5.18.0
pymongo==4.6.1
python-dotenv==1.0.0
sqlparse==0.4.4
tenacity==8.2.3
typing_extensions==4.9.0
tzdata==2023.4
